pgp_alternatives/requirements_gnupg.txt
boto3>=1.26.0
botocore>=1.29.0
python-gnupg>=0.5.0
